third_party/blink/web_tests/platform/generic/fast/frames/frame-focus-send-blur-expected.txt

Make sure that moving focus on an inactive frame fires blur event on existing element focus.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS A focus event was dispatched on frame1 input element.
===> Making a first frame with a focused element inactive by focus() for second frame.
PASS A blur event was dispatched on frame1 input element.
PASS frame1.contentDocument.activeElement is doc.body
===> Making the first frame active again by focus() for first frame.
PASS frame1.contentDocument.activeElement is doc.body
 
